requirements.txt
langchain-google-genai 
crewai 
crewai_tools
python-dotenv
streamlit